refactor: move tables to separate file
